test: set start method to `spawn`
See the following warning:
`DeprecationWarning: This process (pid=578) is multi-threaded,
use of fork() may lead to deadlocks in the child.`

We shouldn't fork child processes when running multiple processes in parallel.
Set the `multiprocessing.set_start_method` to `spawn` and move `port_forward`
out of the class so it can be pickled.

Ticket: QA-1107

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>